+++ Modules/Intent.Modules.Application.FluentValidation.Dtos/Templates/ValidationServiceInterface/ValidationServiceInterfaceTemplatePartial.cs
@@ -1,5 +1,8 @@
 using System.Collections.Generic;
+using System.Linq;
 using Intent.Engine;
+using Intent.Modelers.Services.Api;
+using Intent.Modules.Application.FluentValidation.Templates;
 using Intent.Modules.Common;
 using Intent.Modules.Common.CSharp.Templates;
 using Intent.Modules.Common.Templates;
@@ -9,12 +12,12 @@ using Intent.Templates;
 [assembly: DefaultIntentManaged(Mode.Fully)]
 [assembly: IntentTemplate("Intent.ModuleBuilder.CSharp.Templates.CSharpTemplatePartial", Version = "1.0")]
 
-namespace Intent.Modules.Application.ServiceImplementations.FluentValidation.Templates.ValidationServiceInterface
+namespace Intent.Modules.Application.FluentValidation.Dtos.Templates.ValidationServiceInterface
 {
     [IntentManaged(Mode.Fully, Body = Mode.Merge)]
     partial class ValidationServiceInterfaceTemplate : CSharpTemplateBase<object>
     {
-        public const string TemplateId = "Intent.Application.ServiceImplementations.FluentValidation.ValidationServiceInterface";
+        public const string TemplateId = "Intent.Application.FluentValidation.Dtos.ValidationServiceInterface";
 
         [IntentManaged(Mode.Fully, Body = Mode.Ignore)]
         public ValidationServiceInterfaceTemplate(IOutputTarget outputTarget, object model = null) : base(TemplateId, outputTarget, model)